requirements.txt
pandas
numpy
pyyaml
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
